refactor(sellproduct): 简化优先物品识别流程
